feat: add code block and image buttons to custom toolbar in rich editor
